requirements.txt
python-telegram-bot
google-generativeai
pyyaml
requests
python-dotenv
SQLAlchemy
psycopg
asyncpg